# Improve renewal RPC performance