--- a/.changeset/improve_performance_of_renewing_large_contracts.md
+++ b/.changeset/improve_performance_of_renewing_large_contracts.md
@@ -1,0 +1,5 @@
+---
+default: patch
+---
+
+# Improved performance when renewing large contracts